requirements.txt
streamlit==1.30.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.13
numpy>=1.26.0
pandas>=2.1.0
Pillow>=10.1.0
joblib>=1.3.2
scikit-learn>=1.3.0
plotly>=5.18.0
statsmodels>=0.14.0
tensorflow>=2.15.0
pyyaml>=6.0.1
tqdm>=4.66.1
matplotlib>=3.7.2 